Move common code from mailing list/event objects into base class
