+++ GetIntoTeachingApi/Models/MailingListAddMember.cs
@@ -1,44 +1,15 @@
-﻿using System;
-using System.Linq;
-using System.Text.Json.Serialization;
-using GetIntoTeachingApi.Attributes;
+﻿using System.Text.Json.Serialization;
 using Swashbuckle.AspNetCore.Annotations;
 
 namespace GetIntoTeachingApi.Models
 {
-    [Loggable]
-    public class MailingListAddMember
+    public class MailingListAddMember : AddCandidate
     {
-        public Guid? CandidateId { get; set; }
-        public Guid? QualificationId { get; set; }
-        public Guid? PreferredTeachingSubjectId { get; set; }
-        [SwaggerSchema(WriteOnly = true)]
-        public Guid? AcceptedPolicyId { get; set; }
-
-        public int? ConsiderationJourneyStageId { get; set; }
-        public int? DegreeStatusId { get; set; }
         [SwaggerSchema(WriteOnly = true)]
         public int? ChannelId { get; set; }
 
-        [SensitiveData]
-        public string Email { get; set; }
-        [SensitiveData]
-        public string FirstName { get; set; }
-        [SensitiveData]
-        public string LastName { get; set; }
-        [SensitiveData]
-        public string AddressPostcode { get; set; }
-        [SensitiveData]
-        public string Telephone { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToEvents { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToMailingList { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToTeacherTrainingAdviser { get; set; }
-
         [JsonIgnore]
-        public Candidate Candidate => CreateCandidate();
+        public override Candidate Candidate => CreateCandidate();
 
         public MailingListAddMember()
         {
@@ -49,103 +20,39 @@ namespace GetIntoTeachingApi.Models
             PopulateWithCandidate(candidate);
         }
 
-        private void PopulateWithCandidate(Candidate candidate)
+        protected override Candidate CreateCandidate()
         {
-            var latestQualification = candidate.Qualifications.OrderByDescending(q => q.CreatedAt).FirstOrDefault();
+            var candidate = base.CreateCandidate();
+            candidate.ChannelId = CandidateId == null ? ChannelId ?? (int?)Candidate.Channel.MailingList : null;
 
-            if (latestQualification != null)
-            {
-                QualificationId = latestQualification.Id;
-                DegreeStatusId = latestQualification.DegreeStatusId;
-            }
-
-            CandidateId = candidate.Id;
-            PreferredTeachingSubjectId = candidate.PreferredTeachingSubjectId;
-
-            ConsiderationJourneyStageId = candidate.ConsiderationJourneyStageId;
-
-            Email = candidate.Email;
-            FirstName = candidate.FirstName;
-            LastName = candidate.LastName;
-            AddressPostcode = candidate.AddressPostcode;
-            Telephone = candidate.Telephone;
-
-            AlreadySubscribedToMailingList = candidate.HasMailingListSubscription == true;
-            AlreadySubscribedToEvents = candidate.HasEventsSubscription == true;
-            AlreadySubscribedToTeacherTrainingAdviser = candidate.HasTeacherTrainingAdviser();
-        }
-
-        private Candidate CreateCandidate()
-        {
-            var candidate = new Candidate()
-            {
-                Id = CandidateId,
-                ConsiderationJourneyStageId = ConsiderationJourneyStageId,
-                PreferredTeachingSubjectId = PreferredTeachingSubjectId,
-                Email = Email,
-                FirstName = FirstName,
-                LastName = LastName,
-                AddressPostcode = AddressPostcode,
-                Telephone = Telephone,
-                ChannelId = CandidateId == null ? ChannelId ?? (int?)Candidate.Channel.MailingList : null,
-                OptOutOfSms = false,
-                DoNotBulkEmail = false,
-                DoNotEmail = false,
-                DoNotBulkPostalMail = true,
-                DoNotPostalMail = true,
-                DoNotSendMm = false,
-                EligibilityRulesPassed = "false",
-            };
-
-            ConfigureSubscriptions(candidate);
+            ConfigureConsent(candidate);
             AddQualification(candidate);
             AcceptPrivacyPolicy(candidate);
+            ConfigureSubscriptions(candidate);
 
             return candidate;
         }
 
-        private void AddQualification(Candidate candidate)
+        protected override void ConfigureConsent(Candidate candidate)
         {
-            candidate.Qualifications.Add(new CandidateQualification()
-            {
-                Id = QualificationId,
-                DegreeStatusId = DegreeStatusId,
-                TypeId = (int)CandidateQualification.DegreeType.Degree,
-            });
+            base.ConfigureConsent(candidate);
+
+            candidate.DoNotBulkPostalMail = true;
+            candidate.DoNotPostalMail = true;
+            candidate.DoNotSendMm = false;
+            candidate.EligibilityRulesPassed = "false";
         }
 
         private void ConfigureSubscriptions(Candidate candidate)
         {
-            candidate.HasMailingListSubscription = true;
+            ConfigureMailingListSubscriptions(candidate);
             candidate.MailingListSubscriptionChannelId = ChannelId ?? (int)Candidate.SubscriptionChannel.MailingList;
-            candidate.MailingListSubscriptionStartAt = DateTime.UtcNow;
-            candidate.MailingListSubscriptionDoNotEmail = false;
-            candidate.MailingListSubscriptionDoNotBulkEmail = false;
-            candidate.MailingListSubscriptionDoNotBulkPostalMail = true;
-            candidate.MailingListSubscriptionDoNotPostalMail = true;
-            candidate.MailingListSubscriptionDoNotSendMm = false;
 
-            if (string.IsNullOrWhiteSpace(AddressPostcode))
+            if (!string.IsNullOrWhiteSpace(AddressPostcode))
             {
-                return;
-            }
-
-            candidate.HasEventsSubscription = true;
-            candidate.EventsSubscriptionTypeId = (int)Candidate.SubscriptionType.LocalEvent;
-            candidate.EventsSubscriptionChannelId = ChannelId ?? (int)Candidate.SubscriptionChannel.Events;
-            candidate.EventsSubscriptionStartAt = DateTime.UtcNow;
-            candidate.EventsSubscriptionDoNotEmail = false;
-            candidate.EventsSubscriptionDoNotBulkEmail = false;
-            candidate.EventsSubscriptionDoNotBulkPostalMail = true;
-            candidate.EventsSubscriptionDoNotPostalMail = true;
-            candidate.EventsSubscriptionDoNotSendMm = false;
-        }
-
-        private void AcceptPrivacyPolicy(Candidate candidate)
-        {
-            if (AcceptedPolicyId != null)
-            {
-                candidate.PrivacyPolicy = new CandidatePrivacyPolicy() { AcceptedPolicyId = (Guid)AcceptedPolicyId };
+                ConfigureEventsSubscriptions(candidate);
+                candidate.EventsSubscriptionTypeId = (int)Candidate.SubscriptionType.LocalEvent;
+                candidate.EventsSubscriptionChannelId = ChannelId ?? (int)Candidate.SubscriptionChannel.Events;
             }
         }
     }
