+++ GetIntoTeachingApi/Models/TeachingEventAddAttendee.cs
@@ -1,49 +1,21 @@
 ﻿using System;
-using System.Linq;
 using System.Text.Json.Serialization;
-using GetIntoTeachingApi.Attributes;
 using Swashbuckle.AspNetCore.Annotations;
 
 namespace GetIntoTeachingApi.Models
 {
-    [Loggable]
-    public class TeachingEventAddAttendee
+    public class TeachingEventAddAttendee : AddCandidate
     {
-        public Guid? CandidateId { get; set; }
-        public Guid? QualificationId { get; set; }
-
         [SwaggerSchema(WriteOnly = true)]
         public Guid? EventId { get; set; }
-        [SwaggerSchema(WriteOnly = true)]
-        public Guid? AcceptedPolicyId { get; set; }
-        public Guid? PreferredTeachingSubjectId { get; set; }
 
-        public int? ConsiderationJourneyStageId { get; set; }
-        public int? DegreeStatusId { get; set; }
-
-        [SensitiveData]
-        public string Email { get; set; }
-        [SensitiveData]
-        public string FirstName { get; set; }
-        [SensitiveData]
-        public string LastName { get; set; }
-        [SensitiveData]
-        public string AddressPostcode { get; set; }
-        [SensitiveData]
-        public string Telephone { get; set; }
         [SwaggerSchema(ReadOnly = true)]
         public bool SubscribeToEvents => AddressPostcode != null && SubscribeToMailingList;
         [SwaggerSchema(WriteOnly = true)]
         public bool SubscribeToMailingList { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToEvents { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToMailingList { get; set; }
-        [SwaggerSchema(ReadOnly = true)]
-        public bool AlreadySubscribedToTeacherTrainingAdviser { get; set; }
 
         [JsonIgnore]
-        public Candidate Candidate => CreateCandidate();
+        public override Candidate Candidate => CreateCandidate();
 
         public TeachingEventAddAttendee()
         {
@@ -54,109 +26,44 @@ namespace GetIntoTeachingApi.Models
             PopulateWithCandidate(candidate);
         }
 
-        private void PopulateWithCandidate(Candidate candidate)
+        protected override Candidate CreateCandidate()
         {
-            var latestQualification = candidate.Qualifications.OrderByDescending(q => q.CreatedAt).FirstOrDefault();
+            var candidate = base.CreateCandidate();
+            candidate.ChannelId = CandidateId == null ? (int?)Candidate.Channel.Event : null;
 
-            if (latestQualification != null)
-            {
-                QualificationId = latestQualification.Id;
-                DegreeStatusId = latestQualification.DegreeStatusId;
-            }
-
-            CandidateId = candidate.Id;
-            PreferredTeachingSubjectId = candidate.PreferredTeachingSubjectId;
-
-            ConsiderationJourneyStageId = candidate.ConsiderationJourneyStageId;
-
-            Email = candidate.Email;
-            FirstName = candidate.FirstName;
-            LastName = candidate.LastName;
-            AddressPostcode = candidate.AddressPostcode;
-            Telephone = candidate.Telephone;
-
-            AlreadySubscribedToMailingList = candidate.HasMailingListSubscription == true;
-            AlreadySubscribedToEvents = candidate.HasEventsSubscription == true;
-            AlreadySubscribedToTeacherTrainingAdviser = candidate.HasTeacherTrainingAdviser();
-        }
-
-        private Candidate CreateCandidate()
-        {
-            var candidate = new Candidate()
-            {
-                Id = CandidateId,
-                ConsiderationJourneyStageId = ConsiderationJourneyStageId,
-                PreferredTeachingSubjectId = PreferredTeachingSubjectId,
-                Email = Email,
-                FirstName = FirstName,
-                LastName = LastName,
-                AddressPostcode = AddressPostcode,
-                Telephone = Telephone,
-                ChannelId = CandidateId == null ? (int?)Candidate.Channel.Event : null,
-            };
-
-            AddTeachingEventRegistration(candidate);
+            ConfigureConsent(candidate);
             AddQualification(candidate);
             AcceptPrivacyPolicy(candidate);
             ConfigureSubscriptions(candidate);
-            ConfigureConsent(candidate);
+            AddTeachingEventRegistration(candidate);
 
             return candidate;
         }
 
-        private void AddQualification(Candidate candidate)
+        protected override void ConfigureConsent(Candidate candidate)
         {
-            candidate.Qualifications.Add(new CandidateQualification()
-            {
-                Id = QualificationId,
-                DegreeStatusId = DegreeStatusId,
-                TypeId = (int)CandidateQualification.DegreeType.Degree,
-            });
-        }
+            base.ConfigureConsent(candidate);
 
-        private void ConfigureConsent(Candidate candidate)
-        {
-            candidate.OptOutOfSms = false;
-            candidate.DoNotBulkEmail = false;
             candidate.DoNotBulkPostalMail = !SubscribeToMailingList;
-            candidate.DoNotEmail = false;
             candidate.DoNotPostalMail = !SubscribeToMailingList;
             candidate.DoNotSendMm = !SubscribeToEvents && !SubscribeToMailingList;
         }
 
         private void ConfigureSubscriptions(Candidate candidate)
         {
-            candidate.HasEventsSubscription = true;
+            ConfigureEventsSubscriptions(candidate);
+
             candidate.EventsSubscriptionChannelId = (int)Candidate.SubscriptionChannel.Events;
-            candidate.EventsSubscriptionStartAt = DateTime.UtcNow;
-            candidate.EventsSubscriptionDoNotEmail = false;
-            candidate.EventsSubscriptionDoNotBulkEmail = false;
-            candidate.EventsSubscriptionDoNotBulkPostalMail = true;
-            candidate.EventsSubscriptionDoNotPostalMail = true;
-            candidate.EventsSubscriptionDoNotSendMm = false;
 
-            if (string.IsNullOrWhiteSpace(AddressPostcode))
-            {
-                candidate.EventsSubscriptionTypeId = (int)Candidate.SubscriptionType.SingleEvent;
-            }
-            else
-            {
-                candidate.EventsSubscriptionTypeId = (int)Candidate.SubscriptionType.LocalEvent;
-            }
+            candidate.EventsSubscriptionTypeId = string.IsNullOrWhiteSpace(AddressPostcode)
+                ? (int)Candidate.SubscriptionType.SingleEvent
+                : (int)Candidate.SubscriptionType.LocalEvent;
 
-            if (!SubscribeToMailingList)
+            if (SubscribeToMailingList)
             {
-                return;
+                ConfigureMailingListSubscriptions(candidate);
+                candidate.MailingListSubscriptionChannelId = (int)Candidate.SubscriptionChannel.MailingList;
             }
-
-            candidate.HasMailingListSubscription = true;
-            candidate.MailingListSubscriptionChannelId = (int)Candidate.SubscriptionChannel.MailingList;
-            candidate.MailingListSubscriptionStartAt = DateTime.UtcNow;
-            candidate.MailingListSubscriptionDoNotEmail = false;
-            candidate.MailingListSubscriptionDoNotBulkEmail = false;
-            candidate.MailingListSubscriptionDoNotBulkPostalMail = true;
-            candidate.MailingListSubscriptionDoNotPostalMail = true;
-            candidate.MailingListSubscriptionDoNotSendMm = false;
         }
 
         private void AddTeachingEventRegistration(Candidate candidate)
@@ -170,14 +77,6 @@ namespace GetIntoTeachingApi.Models
                     IsCancelled = false,
                     RegistrationNotificationSeen = false,
                 });
-            }
-        }
-
-        private void AcceptPrivacyPolicy(Candidate candidate)
-        {
-            if (AcceptedPolicyId != null)
-            {
-                candidate.PrivacyPolicy = new CandidatePrivacyPolicy() { AcceptedPolicyId = (Guid)AcceptedPolicyId };
             }
         }
     }
